Organize order's coupon files in separate packages
