+++ WooCommerce/src/test/kotlin/com/woocommerce/android/ui/orders/creation/coupon/CouponValidatorTest.kt
@@ -1,6 +1,7 @@
 package com.woocommerce.android.ui.orders.creation.coupon
 
 import com.woocommerce.android.tools.SelectedSite
+import com.woocommerce.android.ui.orders.creation.coupon.edit.CouponValidator
 import com.woocommerce.android.viewmodel.BaseUnitTest
 import kotlinx.coroutines.ExperimentalCoroutinesApi
 import org.assertj.core.api.Assertions.assertThat
