+++ WooCommerce/src/test/kotlin/com/woocommerce/android/ui/orders/creation/coupon/OrderCreateCouponEditViewModelTest.kt
@@ -2,6 +2,9 @@ package com.woocommerce.android.ui.orders.creation.coupon
 
 import com.woocommerce.android.extensions.isNotNullOrEmpty
 import com.woocommerce.android.initSavedStateHandle
+import com.woocommerce.android.ui.orders.creation.coupon.edit.CouponValidator
+import com.woocommerce.android.ui.orders.creation.coupon.edit.OrderCreateCouponEditFragmentArgs
+import com.woocommerce.android.ui.orders.creation.coupon.edit.OrderCreateCouponEditViewModel
 import com.woocommerce.android.viewmodel.BaseUnitTest
 import junit.framework.TestCase.assertFalse
 import junit.framework.TestCase.assertTrue
@@ -20,7 +23,7 @@ class OrderCreateCouponEditViewModelTest : BaseUnitTest() {
     }
     @Test
     fun `given non empty coupon, when passed to coupon edition screen, then should show remove button`() {
-        val navArgs = OrderCreateCouponEditionFragmentArgs("coupon_code").initSavedStateHandle()
+        val navArgs = OrderCreateCouponEditFragmentArgs("coupon_code").initSavedStateHandle()
 
         val sut = OrderCreateCouponEditViewModel(validator, navArgs)
 
@@ -31,7 +34,7 @@ class OrderCreateCouponEditViewModelTest : BaseUnitTest() {
 
     @Test
     fun `given empty coupon, when passed to coupon edition screen, then should not show remove button`() {
-        val navArgs = OrderCreateCouponEditionFragmentArgs("coupon_code").initSavedStateHandle()
+        val navArgs = OrderCreateCouponEditFragmentArgs("coupon_code").initSavedStateHandle()
 
         val sut = OrderCreateCouponEditViewModel(validator, navArgs)
 
@@ -42,7 +45,7 @@ class OrderCreateCouponEditViewModelTest : BaseUnitTest() {
 
     @Test
     fun `when non empty coupon code entered, then should show done button`() {
-        val navArgs = OrderCreateCouponEditionFragmentArgs().initSavedStateHandle()
+        val navArgs = OrderCreateCouponEditFragmentArgs().initSavedStateHandle()
 
         val sut = OrderCreateCouponEditViewModel(validator, navArgs)
         sut.onCouponCodeChanged("new_code")
@@ -54,7 +57,7 @@ class OrderCreateCouponEditViewModelTest : BaseUnitTest() {
 
     @Test
     fun `when empty coupon code entered, then should not show done button`() {
-        val navArgs = OrderCreateCouponEditionFragmentArgs().initSavedStateHandle()
+        val navArgs = OrderCreateCouponEditFragmentArgs().initSavedStateHandle()
 
         val sut = OrderCreateCouponEditViewModel(validator, navArgs)
         sut.onCouponCodeChanged("new_code")
@@ -67,7 +70,7 @@ class OrderCreateCouponEditViewModelTest : BaseUnitTest() {
     @Test
     fun `given non empty coupon, when coupon removed, then should clear coupon`() {
         val initialCouponCode = "coupon_code"
-        val navArgs = OrderCreateCouponEditionFragmentArgs(initialCouponCode).initSavedStateHandle()
+        val navArgs = OrderCreateCouponEditFragmentArgs(initialCouponCode).initSavedStateHandle()
 
         val sut = OrderCreateCouponEditViewModel(validator, navArgs)
 
@@ -86,7 +89,7 @@ class OrderCreateCouponEditViewModelTest : BaseUnitTest() {
     @Test
     fun `given non empty coupon, when done button is clicked, then should update order with coupon`() = testBlocking {
         val initialCouponCode = "coupon_code"
-        val navArgs = OrderCreateCouponEditionFragmentArgs(initialCouponCode).initSavedStateHandle()
+        val navArgs = OrderCreateCouponEditFragmentArgs(initialCouponCode).initSavedStateHandle()
 
         val sut = OrderCreateCouponEditViewModel(validator, navArgs)
 
@@ -106,7 +109,7 @@ class OrderCreateCouponEditViewModelTest : BaseUnitTest() {
     @Test
     fun `given non empty coupon, when remove button is clicked, then should remove coupon from order`() {
         val initialCouponCode = "coupon_code"
-        val navArgs = OrderCreateCouponEditionFragmentArgs(initialCouponCode).initSavedStateHandle()
+        val navArgs = OrderCreateCouponEditFragmentArgs(initialCouponCode).initSavedStateHandle()
 
         val sut = OrderCreateCouponEditViewModel(validator, navArgs)
 
@@ -125,7 +128,7 @@ class OrderCreateCouponEditViewModelTest : BaseUnitTest() {
     @Test
     fun `given non empty coupon, when code is modified and remove button is clicked, then should remove correct coupon from order`() {
         val initialCouponCode = "coupon_code"
-        val navArgs = OrderCreateCouponEditionFragmentArgs(initialCouponCode).initSavedStateHandle()
+        val navArgs = OrderCreateCouponEditFragmentArgs(initialCouponCode).initSavedStateHandle()
 
         val sut = OrderCreateCouponEditViewModel(validator, navArgs)
 
@@ -143,7 +146,7 @@ class OrderCreateCouponEditViewModelTest : BaseUnitTest() {
 
     @Test
     fun `given invalid coupon, when done clicked, then should show error`() = testBlocking {
-        val navArgs = OrderCreateCouponEditionFragmentArgs().initSavedStateHandle()
+        val navArgs = OrderCreateCouponEditFragmentArgs().initSavedStateHandle()
         val sut = OrderCreateCouponEditViewModel(validator, navArgs)
 
         var latestViewState: OrderCreateCouponEditViewModel.ViewState? = null
@@ -158,7 +161,7 @@ class OrderCreateCouponEditViewModelTest : BaseUnitTest() {
 
     @Test
     fun `given invalid coupon, when done clicked, then should hide done button`() = testBlocking {
-        val navArgs = OrderCreateCouponEditionFragmentArgs().initSavedStateHandle()
+        val navArgs = OrderCreateCouponEditFragmentArgs().initSavedStateHandle()
         val sut = OrderCreateCouponEditViewModel(validator, navArgs)
 
         var latestViewState: OrderCreateCouponEditViewModel.ViewState? = null
@@ -174,7 +177,7 @@ class OrderCreateCouponEditViewModelTest : BaseUnitTest() {
     @Test
     fun `given invalid coupon, when coupon code modified, then should clear error`() = testBlocking {
         // given
-        val navArgs = OrderCreateCouponEditionFragmentArgs().initSavedStateHandle()
+        val navArgs = OrderCreateCouponEditFragmentArgs().initSavedStateHandle()
         val sut = OrderCreateCouponEditViewModel(validator, navArgs)
         var latestViewState: OrderCreateCouponEditViewModel.ViewState? = null
         sut.viewState.observeForever {
@@ -194,7 +197,7 @@ class OrderCreateCouponEditViewModelTest : BaseUnitTest() {
     fun `given invalid coupon, when coupon code modified, then should show done button`() = testBlocking {
         // given
         whenever(validator.isCouponValid(anyString())).thenReturn(false)
-        val navArgs = OrderCreateCouponEditionFragmentArgs().initSavedStateHandle()
+        val navArgs = OrderCreateCouponEditFragmentArgs().initSavedStateHandle()
         val sut = OrderCreateCouponEditViewModel(validator, navArgs)
         var latestViewState: OrderCreateCouponEditViewModel.ViewState? = null
         sut.viewState.observeForever {
